feat: add musical information and restructure layout for crew and orchestra pages
