+++ .env.development
@@ -1,1 +1,2 @@
 PUBLIC_HOMEPAGE_HOST=http://localhost:5001
+PUBLIC_MUSICAL="kein-pardon"
