+++ .env.production
@@ -1,1 +1,2 @@
 PUBLIC_HOMEPAGE_HOST=http://www.jungesmusical.de
+PUBLIC_MUSICAL="kein-pardon"
