+++ src/_styles/components/dot-list.pcss
@@ -8,7 +8,7 @@
 	grid-template-columns: 1fr 1fr;
 	grid-gap: 1rem;
 	position: relative;
-	margin-bottom: var(--s);
+	margin-bottom: var(--xxs);
 	padding-left: 0;
 
 	&::after {
