+++ src/_styles/layout.pcss
@@ -1,5 +1,5 @@
 body {
-	--main-base-width: 800px;
+	--main-base-width: 1200px;
 
 	background-color: var(--c-bg-1);
 	color: var(--c-fg-1);
@@ -13,4 +13,15 @@ body {
 
 section:not(:last-child) {
 	margin-bottom: var(--xxl);
+}
+
+.card {
+  background-color: var(--c-bg-2);
+  border: 1px solid var(--c-bg-3);
+  border-radius: var(--m);
+  padding: var(--xxs) var(--l);
+  margin: var(--m) 0;
+  font-size: var(--fs-xs);
+  text-wrap: balance;
+  text-align: center;
 }
