+++ src/_styles/variables.pcss
@@ -2,7 +2,7 @@
 	/* Sizes: Renard 10 -> 10, 12.5, 16, 20, 25, 31.5, 40
 		========================================================================== */
 
-	--base: 1.5rem; /* Default: 16px */
+	--base: 1.8rem; /* Default: 16px */
 	--base-xxs: 0.625;
 	--base-xs: 0.78125;
 	--base-s: 1;
@@ -102,8 +102,8 @@
 	--c-fg-2: var(--color-palette-gray-20);
 	--c-fg-3: var(--color-palette-gray-40);
 	--c-bg-1: var(--color-palette-black-100);
-	--c-bg-2: var(--color-palette-gray-90);
-	--c-bg-3: var(--color-palette-gray-80);
+	--c-bg-2: var(--color-palette-gray-80);
+	--c-bg-3: var(--color-palette-gray-70);
 	--c-prim-fg-1: var(--color-palette-purple-50);
 	--c-prim-fg-2: var(--color-palette-purple-60);
 	--c-prim-bg-1: var(--color-palette-purple-10);
