+++ src/app.html
@@ -18,6 +18,6 @@
 		%sveltekit.head%
 	</head>
 	<body data-sveltekit-preload-data="hover">
-		<main>%sveltekit.body%</main>
+		%sveltekit.body%
 	</body>
 </html>
