+++ src/routes/+layout.svelte
@@ -2,4 +2,28 @@
 	import '../app.pcss';
 </script>
 
+<style>
+
+.artwork-video {
+  margin: var(--m);
+  padding:0;
+  object-fit: cover;
+  object-position: center;
+  width: calc(100% - 2 * var(--m));
+  max-width: none;
+  aspect-ratio: 21 / 9;
+  border-radius: var(--m);
+}
+
+</style>
+
+<video autoplay loop muted playsinline poster="/artwork-poster.jpg" class="artwork-video">
+  <source src="/artwork.mp4" type="video/mp4" />
+</video>
+
+
+<main>
+
 <slot></slot>
+
+</main>
