+++ src/routes/+page.svelte
@@ -1,2 +1,5 @@
-<h1>Welcome to SvelteKit</h1>
-<p>Visit <a href="https://svelte.dev/docs/kit">svelte.dev/docs/kit</a> to read the documentation</p>
+<ul>
+	<li><a href="cast">Cast</a></li>
+	<li><a href="orchester">Orchester</a></li>
+	<li><a href="crew">Crew</a></li>
+</ul>
