--- src/routes/[musical]/+page.svelte
@@ -1,9 +1,0 @@
-<script lang="ts">
-	import { page } from '$app/stores';
-</script>
-
-<ul>
-	<li><a href="{$page.url}/cast">Cast</a></li>
-	<li><a href="{$page.url}/orchester">Orchester</a></li>
-	<li><a href="{$page.url}/crew">Crew</a></li>
-</ul>
